[master] Corrected person id datatype.

+++ Person.go
@@ -5,7 +5,7 @@ import(
 )
 
 type Person struct {
-  Id int
+  Id int64
   Nickname string `json:"nickname"`
   FirstName sql.NullString `json:"firstName"`
   MiddleName sql.NullString `json:"middleName"`
